+++ src/index.test.ts
@@ -9,10 +9,7 @@ function shouldContainOnlyOneOperation(ast: any, name: string) {
   expect(ast.definitions[0].operation).toBe(name)
 }
 
-function shouldDeclareVariables(
-  def: any,
-  expected: { type: string; nonNull: boolean }[],
-) {
+function shouldDeclareVariables(def: any, expected: { type: string; nonNull: boolean }[]) {
   expect(def.variableDefinitions.length).toBe(1)
   expected.forEach((e, index) => {
     if (e.nonNull) {
@@ -29,15 +26,8 @@ function shouldHaveSelectionNames(def: any, names: string[]) {
   })
 }
 
-function shouldHaveVariable(
-  def: any,
-  variables: any,
-  index: number,
-  value: any,
-) {
-  expect(variables[def.variableDefinitions[index].variable.name.value]).toBe(
-    value,
-  )
+function shouldHaveVariable(def: any, variables: any, index: number, value: any) {
+  expect(variables[def.variableDefinitions[index].variable.name.value]).toBe(value)
 }
 
 describe("GraphQLClient", () => {
@@ -61,34 +51,30 @@ describe("GraphQLClient", () => {
         const ast = parse(query)
 
         shouldContainOnlyOneOperation(ast, "query")
-        shouldDeclareVariables(ast.definitions[0], [
-          { type: "String", nonNull: true },
-        ])
+        shouldDeclareVariables(ast.definitions[0], [{ type: "String", nonNull: true }])
         shouldHaveVariable(ast.definitions[0], variables, 0, "foo")
 
         shouldHaveSelectionNames(ast.definitions[0], ["user", "viewer"])
         const definition = ast.definitions[0] as OperationDefinitionNode
-        const selections = definition.selectionSet.selections as ReadonlyArray<
-          FieldNode
-        >
+        const selections = definition.selectionSet.selections as ReadonlyArray<FieldNode>
         const userAlias = selections[0].alias.value
         const viewerAlias = selections[1].alias.value
 
         resolve({
           data: {
             [userAlias]: {
-              id: "bar",
+              id: "bar"
             },
-            [viewerAlias]: null,
+            [viewerAlias]: null
           },
           errors: [
             {
               message: "baz",
-              path: [viewerAlias, "name"],
-            } as any,
-          ],
+              path: [viewerAlias, "name"]
+            } as any
+          ]
         })
-      },
+      }
     })
 
     const spy1 = jest.fn()
@@ -102,8 +88,8 @@ describe("GraphQLClient", () => {
       .query(query, {
         name: {
           type: "String!",
-          value: "foo",
-        },
+          value: "foo"
+        }
       })
       .then(spy1)
     const spy2 = jest.fn()
@@ -114,21 +100,21 @@ describe("GraphQLClient", () => {
     expect(spy1.mock.calls[0][0]).toEqual({
       data: {
         user: {
-          id: "bar",
-        },
-      },
+          id: "bar"
+        }
+      }
     })
     expect(spy2.mock.calls.length).toBe(1)
     expect(spy2.mock.calls[0][0]).toEqual({
       data: {
-        viewer: null,
+        viewer: null
       },
       errors: [
         {
           message: "baz",
-          path: ["viewer", "name"],
-        },
-      ],
+          path: ["viewer", "name"]
+        }
+      ]
     })
     done()
   })
@@ -138,7 +124,7 @@ describe("GraphQLClient", () => {
       wait: 0,
       handle(query, variables, resolve) {
         resolve({})
-      },
+      }
     })
 
     const spy1 = jest.fn()
@@ -159,7 +145,7 @@ describe("GraphQLClient", () => {
       wait: 0,
       handle: (query, variables, resolve, reject) => {
         reject("hello")
-      },
+      }
     })
 
     const spy1 = jest.fn()